CacheManager lock time, nest case change, tagIndex modify.

+++ TABAnimatedDemo/TABAnimated/Control/TABCollectionAnimated.m
@@ -105,7 +105,7 @@ NSString * const TABViewAnimatedFooterPrefixString = @"tab_footer_";
     obj.cellSizeArray = @[@(cellSize)];
     obj.cellCountArray = @[@(animatedCount)];
     obj.cellIndexArray = @[@(0)];
-    [obj.runIndexDict setValue:@(0) forKey:[NSString stringWithFormat:@"%ld",toIndex]];
+    [obj.runIndexDict setValue:@(0) forKey:[NSString stringWithFormat:@"%ld",(long)toIndex]];
     return obj;
 }
 
@@ -124,7 +124,7 @@ NSString * const TABViewAnimatedFooterPrefixString = @"tab_footer_";
         for (NSInteger i = 0; i < cellClassArray.count; i++) {
             NSInteger index = i;
             NSInteger value = i;
-            [obj.runIndexDict setValue:@(value) forKey:[NSString stringWithFormat:@"%ld",index]];
+            [obj.runIndexDict setValue:@(value) forKey:[NSString stringWithFormat:@"%ld",(long)index]];
             [newIndexArray addObject:@(index)];
         }
         obj.cellIndexArray = newIndexArray.copy;
@@ -133,7 +133,7 @@ NSString * const TABViewAnimatedFooterPrefixString = @"tab_footer_";
         for (NSInteger i = 0; i < indexArray.count; i++) {
             NSInteger index = [indexArray[i] integerValue];
             NSInteger value = i;
-            [obj.runIndexDict setValue:@(value) forKey:[NSString stringWithFormat:@"%ld",index]];
+            [obj.runIndexDict setValue:@(value) forKey:[NSString stringWithFormat:@"%ld",(long)index]];
         }
     }
     return obj;
